+++ html/index.test.ts
@@ -28,7 +28,7 @@ test('server|html_ head_ body_ div_ a_', ()=>{
 			body_(
 				div_({ id: 'div-id' },
 					a_({ href: 'https://github.com/relementjs/all' }, 'relementjs'))
-			)).render(),
+			)).toString(),
 		`<html><head></head><body><div id="div-id"><a href="https://github.com/relementjs/all">relementjs</a></div></body></html>`
 	)
 })