+++ server/index.test.ts
@@ -36,60 +36,72 @@ test('server__fragment__relement', ()=>{
 	equal(server__fragment__relement, { attach, bind_, tags, tagsNS, fragment_, raw_ })
 })
 test('tags', ()=>{
-	equal(div(
-		p('👋Hello'),
-		ul(
-			li('🗺️World'),
-			li(a({ href: 'https://github.com/relementjs/server/' }, '🍦relement')),
-		),
-	).render(),
-	'<div><p>👋Hello</p><ul><li>🗺️World</li><li><a href="https://github.com/relementjs/server/">🍦relement</a></li></ul></div>')
+	equal(
+		div(
+			p('👋Hello'),
+			ul(
+				li('🗺️World'),
+				li(a({ href: 'https://github.com/relementjs/server/' }, '🍦relement')),
+			),
+		).toString(),
+		'<div><p>👋Hello</p><ul><li>🗺️World</li><li><a href="https://github.com/relementjs/server/">🍦relement</a></li></ul></div>')
 })
 test('tags|undefined & null prop', ()=>{
-	equal(div({ id: undefined }).render(), '<div></div>')
-	equal(div({ id: null }).render(), '<div></div>')
+	equal('' + div({ id: undefined }), '<div></div>')
+	equal('' + div({ id: null }), '<div></div>')
 })
 test('tags|prop|data- props', ()=>{
-	equal(div({ 'data-foo': JSON.stringify(prop_data__div_o) }).render(), prop_data__div_html)
+	equal('' + div({ 'data-foo': JSON.stringify(prop_data__div_o) }), prop_data__div_html)
 })
 test('elements without child', ()=>{
-	equal(br().render(), '<br>')
-	equal(hr({ class: 'large' }).render(), '<hr class="large">')
+	equal('' + br(), '<br>')
+	equal('' + hr({ class: 'large' }), '<hr class="large">')
 	// Children are ignored even when they are provided
-	equal(br(div('Line')).render(), '<br>')
+	equal('' + br(div('Line')), '<br>')
 })
 test('boolean prop', ()=>{
-	equal(input({ type: 'checkbox', checked: false }).render(), '<input type="checkbox">')
-	equal(input({ type: 'checkbox', checked: true }).render(), '<input type="checkbox" checked>')
-	equal(input({ checked: false }).render(), '<input>')
-	equal(input({ checked: true }).render(), '<input checked>')
+	equal('' + input({ type: 'checkbox', checked: false }), '<input type="checkbox">')
+	equal('' + input({ type: 'checkbox', checked: true }), '<input type="checkbox" checked>')
+	equal('' + input({ checked: false }), '<input>')
+	equal('' + input({ checked: true }), '<input checked>')
 })
 test('escape', ()=>{
-	equal(p('<input>').render(), '<p>&lt;input&gt;</p>')
-	equal(div('a && b').render(), '<div>a &amp;&amp; b</div>')
-	equal(div('<input a && b>').render(), '<div>&lt;input a &amp;&amp; b&gt;</div>')
+	equal('' + p('<input>'), '<p>&lt;input&gt;</p>')
+	equal('' + div('a && b'), '<div>a &amp;&amp; b</div>')
+	equal('' + div('<input a && b>'), '<div>&lt;input a &amp;&amp; b&gt;</div>')
 })
 test('escapeAttr', ()=>{
-	equal(input({ value: '"text"' }).render(), '<input value="&quot;text&quot;">')
+	equal('' + input({ value: '"text"' }), '<input value="&quot;text&quot;">')
 })
 test('nested children', ()=>{
-	equal(ul([li('Item 1'), li('Item 2'), li('Item 3')]).render(),
+	equal('' + ul([li('Item 1'), li('Item 2'), li('Item 3')]),
 		'<ul><li>Item 1</li><li>Item 2</li><li>Item 3</li></ul>')
 	// Deeply nested
-	equal(ul([[li('Item 1'), [li('Item 2')]], li('Item 3')]).render(),
+	equal('' + ul([[li('Item 1'), [li('Item 2')]], li('Item 3')]),
 		'<ul><li>Item 1</li><li>Item 2</li><li>Item 3</li></ul>')
 })
 test('null or undefined are ignored', ()=>{
-	equal(ul(li('Item 1'), li('Item 2'), undefined, li('Item 3'), null).render(),
+	equal('' + ul(li('Item 1'), li('Item 2'), undefined, li('Item 3'), null),
 		'<ul><li>Item 1</li><li>Item 2</li><li>Item 3</li></ul>')
-	equal(ul([li('Item 1'), li('Item 2'), undefined, li('Item 3'), null]).render(),
+	equal('' + ul([li('Item 1'), li('Item 2'), undefined, li('Item 3'), null]),
 		'<ul><li>Item 1</li><li>Item 2</li><li>Item 3</li></ul>')
 	// Deeply nested
-	equal(ul([[undefined, li('Item 1'), null, [li('Item 2')]], null, li('Item 3'), undefined]).render(),
+	equal('' + ul([[undefined, li('Item 1'), null, [li('Item 2')]], null, li('Item 3'), undefined]),
 		'<ul><li>Item 1</li><li>Item 2</li><li>Item 3</li></ul>')
 })
+test('inside template strings', ()=>{
+	equal(`${
+		div(
+			p('👋Hello'),
+			ul(
+				li('🗺️World'),
+				li(a({ href: 'https://github.com/relementjs/server/' }, '🍦relement')),
+			),)
+	}`,
+	'<div><p>👋Hello</p><ul><li>🗺️World</li><li><a href="https://github.com/relementjs/server/">🍦relement</a></li></ul></div>')
+})
 test('fragment_', ()=>{
-	equal(fragment_(
+	equal('' + fragment_(
 		'<div>Yo!</div>',
 		undefined,
 		null,
@@ -98,65 +110,65 @@ test('fragment_', ()=>{
 			ul(
 				li('🗺️World'),
 				li(a({ href: 'https://github.com/relementjs/server/' }, '🍦relement')),
-			),)).render(),
+			))),
 	'&lt;div&gt;Yo!&lt;/div&gt;<div><p>👋Hello</p><ul><li>🗺️World</li><li><a href="https://github.com/relementjs/server/">🍦relement</a></li></ul></div>')
 })
 test('raw_', ()=>{
-	equal(raw_('<div>row 0</div><div>row 1</div><div>row 2</div>').render(),
+	equal('' + raw_('<div>row 0</div><div>row 1</div><div>row 2</div>'),
 		'<div>row 0</div><div>row 1</div><div>row 2</div>')
-	equal(raw_(undefined).render(), '')
-	equal(raw_(null).render(), '')
+	equal('' + raw_(undefined), '')
+	equal('' + raw_(null), '')
 })
 test('attach|basic', ()=>{
 	const dom = ul()
 	equal(attach(dom, li('Item 1'), li('Item 2')), dom)
-	equal(dom.render(), '<ul><li>Item 1</li><li>Item 2</li></ul>')
+	equal('' + dom, '<ul><li>Item 1</li><li>Item 2</li></ul>')
 	equal(attach(dom, li('Item 3'), li('Item 4'), li('Item 5')), dom)
-	equal(dom.render(), '<ul><li>Item 1</li><li>Item 2</li><li>Item 3</li><li>Item 4</li><li>Item 5</li></ul>')
+	equal('' + dom, '<ul><li>Item 1</li><li>Item 2</li><li>Item 3</li><li>Item 4</li><li>Item 5</li></ul>')
 	// No-op if no children specified
 	equal(attach(dom), dom)
-	equal(dom.render(), '<ul><li>Item 1</li><li>Item 2</li><li>Item 3</li><li>Item 4</li><li>Item 5</li></ul>')
+	equal('' + dom, '<ul><li>Item 1</li><li>Item 2</li><li>Item 3</li><li>Item 4</li><li>Item 5</li></ul>')
 })
 test('attach|nested children', ()=>{
 	const dom = ul()
 	equal(attach(dom, [li('Item 1'), li('Item 2')]), dom)
-	equal(dom.render(), '<ul><li>Item 1</li><li>Item 2</li></ul>')
+	equal('' + dom, '<ul><li>Item 1</li><li>Item 2</li></ul>')
 	// Deeply nested
 	equal(attach(dom, [[li('Item 3'), [li('Item 4')]], li('Item 5')]), dom)
-	equal(dom.render(), '<ul><li>Item 1</li><li>Item 2</li><li>Item 3</li><li>Item 4</li><li>Item 5</li></ul>')
+	equal('' + dom, '<ul><li>Item 1</li><li>Item 2</li><li>Item 3</li><li>Item 4</li><li>Item 5</li></ul>')
 	// No-op if no children specified
 	equal(attach(dom, [[[]]]), dom)
-	equal(dom.render(), '<ul><li>Item 1</li><li>Item 2</li><li>Item 3</li><li>Item 4</li><li>Item 5</li></ul>')
+	equal('' + dom, '<ul><li>Item 1</li><li>Item 2</li><li>Item 3</li><li>Item 4</li><li>Item 5</li></ul>')
 })
 test('attach|null or undefined are ignored', ()=>{
 	const dom = ul()
 	equal(attach(dom, li('Item 1'), li('Item 2'), undefined, li('Item 3'), null), dom)
-	equal(dom.render(), '<ul><li>Item 1</li><li>Item 2</li><li>Item 3</li></ul>')
+	equal('' + dom, '<ul><li>Item 1</li><li>Item 2</li><li>Item 3</li></ul>')
 	equal(attach(dom, [li('Item 4'), li('Item 5'), undefined, li('Item 6'), null]), dom)
-	equal(dom.render(),
+	equal('' + dom,
 		'<ul><li>Item 1</li><li>Item 2</li><li>Item 3</li><li>Item 4</li><li>Item 5</li><li>Item 6</li></ul>')
 	// Deeply nested
 	equal(attach(dom, [[undefined, li('Item 7'), null, [li('Item 8')]], null, li('Item 9'), undefined]), dom)
-	equal(dom.render(),
+	equal('' + dom,
 		'<ul><li>Item 1</li><li>Item 2</li><li>Item 3</li><li>Item 4</li><li>Item 5</li><li>Item 6</li><li>Item 7</li><li>Item 8</li><li>Item 9</li></ul>')
 	// No-op if no non-empty children specified
 	equal(attach(dom, [[null, [undefined]], undefined], null), dom)
-	equal(dom.render(),
+	equal('' + dom,
 		'<ul><li>Item 1</li><li>Item 2</li><li>Item 3</li><li>Item 4</li><li>Item 5</li><li>Item 6</li><li>Item 7</li><li>Item 8</li><li>Item 9</li></ul>')
 })
 test('onclick handler', ()=>{
 	{
 		const dom = div(button({ onclick: 'alert("Hello")' }, 'Click me'))
-		equal(dom.render(), '<div><button onclick="alert(&quot;Hello&quot;)">Click me</button></div>')
+		equal('' + dom, '<div><button onclick="alert(&quot;Hello&quot;)">Click me</button></div>')
 	}
 	// {
 	// 	const dom = div(button({ onClick: 'alert("Hello")' }, 'Click me'))
-	// 	equal(dom.render(), '<div><button onclick="alert(&quot;Hello&quot;)">Click me</button></div>')
+	// 	equal('' + dom, '<div><button onclick="alert(&quot;Hello&quot;)">Click me</button></div>')
 	// }
 	{
 		// Function-valued onclick handler will be skipped
 		const dom = div(button({ onclick: ()=>alert('Hello') }, 'Click me'))
-		equal(dom.render(), '<div><button>Click me</button></div>')
+		equal('' + dom, '<div><button>Click me</button></div>')
 	}
 })
 test('tagsNS: svg', ()=>{
@@ -171,7 +183,7 @@ test('tagsNS: svg', ()=>{
 		circle({ cx: '34', cy: '20', 'r': '2', stroke: 'black', 'stroke-width': '2', fill: 'black' }),
 		path({ 'd': 'M 15 30 Q 25 40, 35 30', stroke: 'black', 'stroke-width': '2', fill: 'transparent' }),
 	)
-	equal(dom.render(),
+	equal('' + dom,
 		'<svg width="16px" viewbox="0 0 50 50"><circle cx="25" cy="25" r="20" stroke="black" stroke-width="2" fill="yellow"></circle><circle cx="16" cy="20" r="2" stroke="black" stroke-width="2" fill="black"></circle><circle cx="34" cy="20" r="2" stroke="black" stroke-width="2" fill="black"></circle><path d="M 15 30 Q 25 40, 35 30" stroke="black" stroke-width="2" fill="transparent"></path></svg>')
 })
 test('tagsNS: math', ()=>{
@@ -184,7 +196,7 @@ test('tagsNS: math', ()=>{
 		msup
 	} = tagsNS<'mathml'>('http://www.w3.org/1998/Math/MathML')
 	const dom = math(msup(mi('e'), mrow(mi('i'), mi('π'))), mo('+'), mn('1'), mo('='), mn('0'))
-	equal(dom.render(),
+	equal('' + dom,
 		'<math><msup><mi>e</mi><mrow><mi>i</mi><mi>π</mi></mrow></msup><mo>+</mo><mn>1</mn><mo>=</mo><mn>0</mn></math>')
 })
 test('reactive', ()=>{
@@ -214,7 +226,7 @@ test('reactive', ()=>{
 		),
 		()=>(state5() ? pre : div)(state3),
 	)
-	equal(dom.render(),
+	equal('' + dom,
 		'<div>1<span>2</span><p>Prefix - abc</p>- Suffix<p data-index="1" data-id="4" data-title="abc" data-text="Prefix - Suffix">1abcabcabc</p><button onclick="alert(&quot;Hello&quot;)">Button1</button><button>Button2</button><div>abc</div></div>')
 })
 test('html_', ()=>{
@@ -230,10 +242,10 @@ test('html_', ()=>{
 // Test cases for examples used in the documentation. Having the tests to ensure the examples
 // are always correct.
 test('example: van-plate-server', ()=>{
-	equal(a({ href: 'https://github.com/relementjs/server/' }, '🍦relement').render(),
+	equal('' + a({ href: 'https://github.com/relementjs/server/' }, '🍦relement'),
 		`<a href="https://github.com/relementjs/server/">🍦relement</a>`)
-	equal(button({ onclick: 'alert("Hello")' }, 'Click').render(),
+	equal('' + button({ onclick: 'alert("Hello")' }, 'Click'),
 		`<button onclick="alert(&quot;Hello&quot;)">Click</button>`)
-	equal(input({ type: 'text', value: 'value' }).render(), `<input type="text" value="value">`)
+	equal('' + input({ type: 'text', value: 'value' }), `<input type="text" value="value">`)
 })
 test.run()